Set beat length of x/8 to 1/4
Empirically, it fixes
https://github.com/tobyspark/folk-rnn-webapp/issues/47
https://github.com/paulrosen/abcjs/issues/155

Why was it 3/8? Because these are compound meter where each beat contains a triple pulse? This could just be a silly fix, or there might be some inconsistency around what's a beat across the abcjs code base, and this works for now, until one of those other parts is touched.